Token Validation

+++ src/files/files.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, Post, Put, Delete, Patch, Body, Param, Req, RawBodyReq
 import { FilesService } from './files.service';
 import { ValidationPipe, ParseUUIDPipe, UploadedFile, UseInterceptors, UploadedFiles } from '@nestjs/common';
 import { FileInterceptor, FilesInterceptor } from '@nestjs/platform-express';
-import { imageFileFilter } from '../utils/file-upload.utils';
+import { imageFileFilter, getToken } from '../utils/file-upload.utils';
 import { diskStorage } from 'multer';
 const fs = require('fs');
 
@@ -11,14 +11,6 @@ export class FilesController {
   constructor(
     private readonly filesService: FilesService
   ) { }
-
-  // @Post('upload')
-  // @UseInterceptors(FilesInterceptor('file', 1, { dest: './data/tmp' }))
-  // async uploadFile(@UploadedFiles() file, @Body() body) {
-  //   console.log(body.name);
-  //   console.log('file----', file);
-  //   return this.filesService.uploadFile(file);
-  // }
 
   @Post('upload')
   @UseInterceptors(
@@ -37,43 +29,57 @@ export class FilesController {
       fileFilter: imageFileFilter,
     }),
   )
-  async uploadedFile(@UploadedFile() file) {
-    // console.log('uploadedFile----------------------------');
-    // console.log(file);
-    if (file.originalname !== 'tmpImage.webp') {
-      this.filesService.deleteFile('data/tmp/' + file.originalname);
+  async uploadedFile(@Req() request: RawBodyRequest<Request>, @UploadedFile() file) {
+    const Token = await getToken(request.headers);
+    const validateToken = await this.filesService.validateToken(Token);
+    if (validateToken) {
+      if (file.originalname !== 'tmpImage.webp') {
+        this.filesService.deleteFile('data/tmp/' + file.originalname);
+      }
+      const response = {
+        originalname: file.originalname,
+        filename: file.filename,
+        src: `http://localhost:5000/tmp/${file.filename}`
+      };
+      return response;
     }
-    const response = {
-      originalname: file.originalname,
-      filename: file.filename,
-      src: `http://localhost:5000/tmp/${file.filename}`
-    };
-    return response;
   }
 
   @Delete()
-  delete(@Req() req: RawBodyRequest<Request>) {
-    return this.filesService.deleteFile(req.body);
+  async delete(@Req() request: RawBodyRequest<Request>) {
+    const Token = await getToken(request.headers);
+    const validateToken = await this.filesService.validateToken(Token);
+    if (validateToken) {
+      return this.filesService.deleteFile(request.body);
+    }
   }
 
   @Delete('tmp/:id')
-  deleteTmp(@Param('id') id): Promise<any> {
-    return this.filesService.deleteFile({ fileName: 'data/tmp/' + id });
+  async deleteTmp(@Req() request: RawBodyRequest<Request>, @Param('id') id): Promise<any> {
+    const Token = await getToken(request.headers);
+    const validateToken = await this.filesService.validateToken(Token)
+    if (validateToken) {
+      return this.filesService.deleteFile({ fileName: 'data/tmp/' + id });
+    }
+
   }
 
   @Patch('moveimages')
-  moveImages(@Req() req: RawBodyRequest<Request>): Promise<any> {
-    return this.filesService.moveImages(req.body);
+  async moveImages(@Req() request: RawBodyRequest<Request>): Promise<any> {
+    const Token = await getToken(request.headers);
+    const validateToken = await this.filesService.validateToken(Token);
+    if (validateToken) {
+      return this.filesService.moveImages(request.body);
+    }
   }
 
   @Patch('json')
-  writejson(@Req() req: RawBodyRequest<Request>) {
-    return this.filesService.saveJson(req.body);
+  async writejson(@Req() request: RawBodyRequest<Request>) {
+    const Token = await getToken(request.headers);
+    const validateToken = await this.filesService.validateToken(Token);
+    if (validateToken) {
+      return this.filesService.saveJson(request.body);
+    }
   }
-
-  // @Patch('writejson')
-  // writejson(@Body() payload: any): Promise<any> {
-  //   return this.filesService.writejson(payload);
-  // }
 
 }
